+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Major changes
 
 - add demo data and organize examples
+- add Relative Log Abundance (RLA) plots and Relative Log Abundance Ridge(RLA) plots
 - add getcsv to save the list as csv file
 - add getfilter to filter the list
 - add getpower to deal with power analysis in metabolomics